Moving methods from spec into the driver
