+++ src/components/tree-view/index.ts
@@ -1,1 +1,2 @@
 export * from './tree-view';
+export * from './tree-view-driver';
